refactor(bookmarks): code refactoring accross bookmarks manager /dashboard
